tools_QGIS/dbQueryMCPLayer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "db-query-mcp-layer"
version = "0.1.0"
description = "Read-only MCP layer for Databricks and PostgreSQL with macOS Keychain-backed secrets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "databricks-sql-connector>=3.2.0",
    "psycopg[binary]>=3.2.0",
]

[project.scripts]
db-query-mcp = "db_query_mcp.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
